Refactor hooks + apply refactor to Discover.js
